+++ packages/editor/components/SideMenu/messages.jsx
@@ -2,7 +2,5 @@ import { defineMessages } from 'react-intl';
 
 export default defineMessages({
   collapse: 'Collapse',
-  editor: 'Editor',
   general: 'General',
-  resources: 'Resources',
 });