+++ packages/editor/components/Toolbar/Toolbar.jsx
@@ -1,9 +1,9 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import { FormattedMessage } from 'react-intl';
 import { Link } from 'react-router-dom';
 
 import { version } from '../../package.json';
+import ProfileDropdown from '../ProfileDropdown';
 import styles from './Toolbar.css';
 import messages from './messages';
 
@@ -11,11 +11,10 @@ export default class Toolbar extends React.Component {
   static propTypes = {
     intl: PropTypes.shape().isRequired,
     isLoggedIn: PropTypes.bool.isRequired,
-    logout: PropTypes.func.isRequired,
   };
 
   render() {
-    const { intl, isLoggedIn, logout } = this.props;
+    const { intl, isLoggedIn } = this.props;
 
     return (
       <nav className={`navbar is-fixed-top is-dark ${styles.root}`}>
@@ -40,29 +39,10 @@ export default class Toolbar extends React.Component {
           </a>
         </div>
         <div className="navbar-brand">
-          <span className="navbar-item">
-            <a
-              className="button"
-              href="https://appsemble.dev"
-              rel="noopener noreferrer"
-              target="_blank"
-            >
-              <FormattedMessage {...messages.docsLink} />
-            </a>
-          </span>
-        </div>
-        <div className="navbar-brand">
           {isLoggedIn && (
-            <span className="navbar-item">
-              <button className="button" onClick={logout} type="button">
-                <span className="icon">
-                  <i className="fas fa-sign-out-alt" />
-                </span>
-                <span>
-                  <FormattedMessage {...messages.logoutButton} />
-                </span>
-              </button>
-            </span>
+            <div className="navbar-item">
+              <ProfileDropdown />
+            </div>
           )}
         </div>
       </nav>