+++ packages/editor/components/Toolbar/messages.jsx
@@ -2,6 +2,4 @@ import { defineMessages } from 'react-intl';
 
 export default defineMessages({
   iconAlt: 'Appsemble logo',
-  docsLink: 'Documentation',
-  logoutButton: 'Logout',
 });