+++ packages/server/api/components/requestBodies/index.js
@@ -1,4 +1,5 @@
 import blockDefinition from './blockDefinition';
 import resource from './resource';
+import user from './user';
 
-export default { blockDefinition, resource };
+export default { blockDefinition, resource, user };