+++ packages/server/api/components/requestBodies/user.js
@@ -1,0 +1,19 @@
+export default {
+  description: 'A user profile',
+  content: {
+    'application/json': {
+      schema: {
+        type: 'object',
+        properties: {
+          name: {
+            $ref: '#/components/schemas/User/properties/name',
+          },
+          primaryEmail: {
+            $ref: '#/components/schemas/User/properties/primaryEmail',
+          },
+        },
+        additionalProperties: true,
+      },
+    },
+  },
+};