+++ packages/server/api/paths/user.js
@@ -1,0 +1,103 @@
+export default {
+  '/api/user': {
+    get: {
+      tags: ['template'],
+      description: "Fetch the logged in user's profile.",
+      operationId: 'getUser',
+      responses: {
+        200: {
+          description: "The user's profile.",
+          content: {
+            'application/json': {
+              schema: {
+                $ref: '#/components/schemas/User',
+              },
+            },
+          },
+        },
+      },
+      security: [{ apiUser: [] }],
+    },
+    put: {
+      tags: ['template'],
+      description: "Update the logged in user's profile.",
+      operationId: 'updateUser',
+      requestBody: {
+        required: true,
+        $ref: '#/components/requestBodies/user',
+      },
+      responses: {
+        200: {
+          description: "The user's profile.",
+          content: {
+            'application/json': {
+              schema: {
+                type: 'object',
+                $ref: '#/components/schemas/User',
+              },
+            },
+          },
+        },
+      },
+      security: [{ apiUser: [] }],
+    },
+  },
+  '/api/user/email': {
+    post: {
+      tags: ['template'],
+      description: "Register a new email to logged in user's account.",
+      operationId: 'addEmail',
+      requestBody: {
+        required: true,
+        content: {
+          'application/json': {
+            schema: {
+              type: 'object',
+              required: ['email'],
+              properties: {
+                email: {
+                  type: 'string',
+                  format: 'email',
+                },
+              },
+            },
+          },
+        },
+      },
+      responses: {
+        201: {
+          description: 'The email address has been added succesfully.',
+        },
+      },
+      security: [{ apiUser: [] }],
+    },
+    delete: {
+      tags: ['template'],
+      description: "Remove an existing email to logged in user's account.",
+      operationId: 'removeEmail',
+      requestBody: {
+        required: true,
+        content: {
+          'application/json': {
+            schema: {
+              type: 'object',
+              required: ['email'],
+              properties: {
+                email: {
+                  type: 'string',
+                  format: 'email',
+                },
+              },
+            },
+          },
+        },
+      },
+      responses: {
+        204: {
+          description: 'The email address has been removed succesfully.',
+        },
+      },
+      security: [{ apiUser: [] }],
+    },
+  },
+};