+++ packages/server/api/tags/index.js
@@ -3,5 +3,6 @@ import asset from './asset';
 import auth from './auth';
 import organization from './organization';
 import resource from './resource';
+import user from './user';
 
-export default [app, asset, auth, organization, resource];
+export default [app, asset, auth, organization, resource, user];