+++ packages/server/migrations/index.js
@@ -1,3 +1,4 @@
 import initial from './0.6.0-initial';
+import user from './0.7.0-user-profile';
 
 export default [initial];