+++ packages/server/utils/email.js
@@ -75,6 +75,19 @@ export async function sendWelcomeEmail({ email, name, url }, smtp) {
   return sendEmail({ to, subject }, content, smtp);
 }
 
+export async function sendAddedEmail({ email, name, url }, smtp) {
+  const replacements = {
+    greeting: name ? `Hello ${name}` : 'Hello',
+    url,
+  };
+
+  const { attributes, content } = processTemplate(readTemplate('emailAdded'), replacements);
+  const { subject } = attributes;
+  const to = name ? `"${name}" <${email}>` : email;
+
+  return sendEmail({ to, subject }, content, smtp);
+}
+
 export async function resendVerificationEmail({ email, name, url }, smtp) {
   const replacements = {
     greeting: name ? `Hello ${name}` : 'Hello',