+++ CHANGELOG.md
@@ -13,6 +13,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Helm: Add official Helm chart.
 - Server: Add support for database migrations.
 
+### Fixed
+
+- Server: Internal server error responses are now logged as such.
+
 ## [0.6.0] - 2019-05-20
 
 ### Added